Created the view pager separate
