+++ app/src/main/java/com/sixamigos/sjsucanvasapp/home/HomeActivity.java
@@ -3,18 +3,12 @@ package com.sixamigos.sjsucanvasapp.home;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 
 import com.sixamigos.sjsucanvasapp.R;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class HomeActivity extends AppCompatActivity {
 
@@ -46,39 +40,10 @@ public class HomeActivity extends AppCompatActivity {
   }
 
   private void setupViewPager(ViewPager viewPager) {
-    Adapter adapter = new Adapter(getSupportFragmentManager());
+    HomePagerAdapter adapter = new HomePagerAdapter(getSupportFragmentManager());
     adapter.addFragment(new HomeScreenFragment(), "Category 1");
     adapter.addFragment(new HomeScreenFragment(), "Category 2");
     adapter.addFragment(new HomeScreenFragment(), "Category 3");
     viewPager.setAdapter(adapter);
-  }
-
-  static class Adapter extends FragmentPagerAdapter {
-    private final List<Fragment> mFragments = new ArrayList<>();
-    private final List<String> mFragmentTitles = new ArrayList<>();
-
-    public Adapter(FragmentManager fm) {
-      super(fm);
-    }
-
-    public void addFragment(Fragment fragment, String title) {
-      mFragments.add(fragment);
-      mFragmentTitles.add(title);
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-      return mFragments.get(position);
-    }
-
-    @Override
-    public int getCount() {
-      return mFragments.size();
-    }
-
-    @Override
-    public CharSequence getPageTitle(int position) {
-      return mFragmentTitles.get(position);
-    }
   }
 }
